h2の依存関係が解決できなくなっていたので対処した。
git-svn-id: https://www.seasar.org/svn/sandbox/cms/trunk@1057 ce2cb714-bd0d-0410-8c66-e472ade7ad34

+++ beantable/src/test/java/org/seasar/cms/beantable/impl/BeantableDaoBaseTest.java
@@ -195,7 +195,7 @@ public class BeantableDaoBaseTest extends BeantableDaoTestCase<Hoge> {
     }
 
     public void testExecute_返り値がNumberでも正しく動作することTx() throws Exception {
-        assertEquals(0, hogeDao_.getDtoCount());
+        assertEquals(0, hogeDao_.getDtoCount().intValue());
     }
 
     public void testExecute_返り値がNumberの配列でも正しく動作することTx() throws Exception {
